New RepeatMasker version, new format
Because the format of RepBase has changed I had to revert back
to the infamous file to feed in to the configure script.
postinstall should work fine but if something goes wrong, it
will die after 10 minutes. It could probably be shorted.
If RepBase is remove, postinstall needs to be rerun to refresh
the repeat library.